backend/pyproject.toml

[project]
name = "myway-backend"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "fastapi==0.111.0",
    "uvicorn[standard]==0.30.0",
    "pydantic==2.8.2",
    "httpx==0.27.0",
    "python-dotenv==1.0.1",
    "google-cloud-firestore==2.16.0",
    "google-cloud-storage==2.16.0",
    "google-cloud-pubsub==2.21.1",
    "google-cloud-aiplatform==1.64.0",
    "respx==0.21.1",
    "Pillow>=10.0.0"
]

[project.optional-dependencies]
test = [
    "pytest==8.3.2",
    "pytest-asyncio==0.23.7",
    "pytest-mock==3.14.0",
    "anyio==4.4.0"
]

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra"
testpaths = ["src/tests"]

[tool.pdm]
